allocation of modules
