+++ lib/chain/src/ChainStore/index.js
@@ -9,20 +9,30 @@ import Immutable from "immutable";
 import {Apis} from "bitsharesjs-ws";
 import ChainValidation from "../ChainValidation";
 import BigInteger from "bigi";
-import ee from "../EmitterInstance";
 import {
+    emitter,
     getObjectType,
+    clearObjectCache,
+    getHeadBlockDate,
     timeStringToDate,
-    DEFAULT_AUTO_SUBSCRIBE,
-    OP_HISTORY,
+    getEstimatedChainTimeOffset,
+    DEBUG,
+    ACCOUNT_PREFIX,
     WITNESS_PREFIX,
     COMMITTEE_PREFIX,
-    ACCOUNT_PREFIX,
-    DEBUG
+    DEFAULT_AUTO_SUBSCRIBE
 } from "./utils";
 
+import {addProposalData} from "./proposal";
 import {
-    subscribers,
+    getAccountRefsOfKey,
+    getAccountRefsOfAccount,
+    getAccountMemberStatus,
+    getAccountBalance,
+    fetchRecentHistory
+} from "./account";
+
+import {
     clearSubscirbers,
     subscribed,
     chainTimeOffset,
@@ -31,28 +41,28 @@ import {
     objectById,
     accountsByName,
     assetsBySymbol,
-    accountIdsByKey,
-    accountIdsByAccount,
     balanceObjectsByAddress,
-    getAccountRefsOfKeysCalls,
-    getAccountRefsOfAccountsCalls,
-    accountHistoryRequests,
     witnessByAccountId,
     workers,
     committeeByAccountId,
     objectsByVoteId,
     fetchingGetFullAccounts,
     getFullAccountsSubscriptions,
-    timeout,
-    dispatched,
     subError,
     headBlockTimeString,
     clearCache
 } from "./store";
 
-let emitter = ee();
+import {
+    _subTo,
+    _isSubbedTo,
+    unSubFrom,
+    subscribe,
+    unsubscribe,
+    notifySubscribers
+} from "./subscribe";
 
-clearCache();
+//clearCache();
 
 const resetCache = subscribe_to_new => {
     subscribed.set(false);
@@ -149,17 +159,6 @@ const init = (subscribe_to_new = true) =>
 
         initSubscribe();
     });
-
-const _subTo = (type, id) => {
-    !subbed[type].has(id) && subbed[type].add(id);
-};
-
-const unSubFrom = (type, id) => {
-    subbed[type].delete(id);
-    objectById.delete(id);
-};
-
-const _isSubbedTo = (type, id) => subbed[type].has(id);
 
 const onUpdate = (
     updated_objects /// map from account id to objects
@@ -259,47 +258,6 @@ const onUpdate = (
     notifySubscribers();
 };
 
-const notifySubscribers = () => {
-    // Dispatch at most only once every x milliseconds
-    if (!dispatched.get()) {
-        dispatched.set(true);
-        timeout.set(
-            setTimeout(() => {
-                dispatched.set(false);
-                subscribers.forEach(callback => {
-                    callback();
-                });
-            }, dispatchFrequency.get())
-        );
-    }
-};
-
-/**
- *  Add a callback that will be called anytime any object in the cache is updated
- */
-const subscribe = callback => {
-    if (subscribers.has(callback))
-        return console.error("Subscribe callback already exists", callback);
-    subscribers.add(callback);
-};
-
-/**
- *  Remove a callback that was previously added via subscribe
- */
-const unsubscribe = callback => {
-    if (!subscribers.has(callback))
-        return console.error("Unsubscribe callback does not exists", callback);
-    subscribers.delete(callback);
-};
-
-/** Clear an object from the cache to force it to be fetched again. This may
- * be useful if a query failed the first time and the wallet has reason to believe
- * it may succeede the second time.
- */
-const clearObjectCache = id => {
-    objectById.delete(id);
-};
-
 /**
  * There are three states an object id could be in:
  *
@@ -391,84 +349,6 @@ const getAsset = id_or_symbol => {
             assetsBySymbol.delete(id_or_symbol);
         });
 
-    return undefined;
-};
-
-/**
- *  @param the public key to find accounts that reference it
- *
- *  @return Set of account ids that reference the given key
- *  @return a empty Set if no items are found
- *  @return undefined if the result is unknown
- *
- *  If this method returns undefined, then it will send a request to
- *  the server for the current set of accounts after which the
- *  server will notify us of any accounts that reference these keys
- */
-const getAccountRefsOfKey = key => {
-    if (getAccountRefsOfKeysCalls.has(key)) return accountIdsByKey.get(key);
-    else {
-        getAccountRefsOfKeysCalls.add(key);
-
-        Apis.db
-            .get_key_references([key])
-            .then(vec_account_id => {
-                let refs = Immutable.Set();
-                vec_account_id = vec_account_id[0];
-                refs = refs.withMutations(r => {
-                    for (let i = 0; i < vec_account_id.length; ++i) {
-                        r.add(vec_account_id[i]);
-                    }
-                });
-                accountIdsByKey.set(key, refs);
-                notifySubscribers();
-            })
-            .catch(err => {
-                console.error("get_key_references", err);
-                accountIdsByKey.delete(key);
-                getAccountRefsOfKeysCalls.delete(key);
-            });
-        return undefined;
-    }
-    //return undefined;
-};
-
-/**
- *  @param the account id to find accounts that reference it
- *
- *  @return Set of account ids that reference the given key
- *  @return a empty Set if no items are found
- *  @return undefined if the result is unknown
- *
- *  If this method returns undefined, then it will send a request to
- *  the server for the current set of accounts after which the
- *  server will notify us of any accounts that reference these keys
- */
-const getAccountRefsOfAccount = account_id => {
-    if (getAccountRefsOfAccountsCalls.has(account_id))
-        return accountIdsByAccount.get(account_id);
-    else {
-        getAccountRefsOfAccountsCalls.add(account_id);
-
-        Apis.db
-            .get_account_references(account_id)
-            .then(vec_account_id => {
-                let refs = Immutable.Set();
-                refs = refs.withMutations(r => {
-                    for (let i = 0; i < vec_account_id.length; ++i) {
-                        r.add(vec_account_id[i]);
-                    }
-                });
-                accountIdsByAccount.set(account_id, refs);
-                notifySubscribers();
-            })
-            .catch(err => {
-                console.error("get_account_references", err);
-                accountIdsByAccount.delete(account_id);
-                getAccountRefsOfAccountsCalls.delete(account_id);
-            });
-        return undefined;
-    }
     return undefined;
 };
 
@@ -942,107 +822,6 @@ const fetchFullAccount = (name_or_id, autosubscribe = null) => {
     return undefined;
 };
 
-const getAccountMemberStatus = account => {
-    if (account === undefined) return undefined;
-    if (account === null) return "unknown";
-    if (account.get("lifetime_referrer") == account.get("id"))
-        return "lifetime";
-    let exp = new Date(account.get("membership_expiration_date")).getTime();
-    let now = new Date().getTime();
-    if (exp < now) return "basic";
-    return "annual";
-};
-
-const getAccountBalance = (account, asset_type) => {
-    let balances = account.get("balances");
-    if (!balances) return 0;
-
-    let balance_obj_id = balances.get(asset_type);
-    if (balance_obj_id) {
-        let bal_obj = objectById.get(balance_obj_id);
-        if (bal_obj) return bal_obj.get("balance");
-    }
-    return 0;
-};
-
-/**
- * There are two ways to extend the account history, add new more
- * recent history, and extend historic hstory. This method will fetch
- * the most recent account history and prepend it to the list of
- * historic operations.
- *
- *  @param account immutable account object
- *  @return a promise with the account history
- */
-const fetchRecentHistory = (account, limit = 100) => {
-    // console.log( "get account history: ", account )
-    /// TODO: make sure we do not submit a query if there is already one
-    /// in flight...
-    let account_id = account;
-    if (!ChainValidation.is_object_id(account_id) && account.toJS)
-        account_id = account.get("id");
-
-    if (!ChainValidation.is_object_id(account_id)) return;
-
-    account = objectById.get(account_id);
-    if (!account || account === true) return;
-
-    let pending_request = accountHistoryRequests.get(account_id);
-    if (pending_request) {
-        pending_request.requests++;
-        return pending_request.promise;
-    } else pending_request = {requests: 0};
-
-    let most_recent = "1." + OP_HISTORY + ".0";
-    let history = account.get("history");
-
-    if (history && history.size) most_recent = history.first().get("id");
-
-    /// starting at 0 means start at NOW, set this to something other than 0
-    /// to skip recent transactions and fetch the tail
-    let start = "1." + OP_HISTORY + ".0";
-
-    pending_request.promise = new Promise((resolve, reject) => {
-        Apis.history
-            .get_account_history(account_id, most_recent, limit, start)
-            .then(operations => {
-                let current_account = objectById.get(account_id);
-                if (typeof current_account !== "object") return;
-                //console.log("current_account", current_account);
-                let current_history = current_account.get("history");
-                if (!current_history) current_history = Immutable.List();
-                let updated_history = Immutable.fromJS(operations);
-                updated_history = updated_history.withMutations(list => {
-                    for (let i = 0; i < current_history.size; ++i)
-                        list.push(current_history.get(i));
-                });
-                let updated_account = current_account.set(
-                    "history",
-                    updated_history
-                );
-                objectById.set(account_id, updated_account);
-
-                //if( current_history != updated_history )
-                //   _notifyAccountSubscribers( account_id )
-
-                let pending_request = accountHistoryRequests.get(account_id);
-                accountHistoryRequests.delete(account_id);
-                if (pending_request.requests > 0) {
-                    // it looks like some more history may have come in while we were
-                    // waiting on the result, lets fetch anything new before we resolve
-                    // this query.
-                    fetchRecentHistory(updated_account, limit).then(
-                        resolve,
-                        reject
-                    );
-                } else resolve(updated_account);
-            }); // end then
-    });
-
-    accountHistoryRequests.set(account_id, pending_request);
-    return pending_request.promise;
-};
-
 /**
  *  Updates the object in place by only merging the set
  *  properties of object.
@@ -1394,39 +1173,6 @@ const getObjectByVoteID = vote_id => {
     return undefined;
 };
 
-const getHeadBlockDate = () => {
-    return timeStringToDate(headBlockTimeString.get());
-};
-
-const getEstimatedChainTimeOffset = () => {
-    if (chainTimeOffset.length === 0) return 0;
-    // Immutable is fast, sorts numbers correctly, and leaves the original unmodified
-    // This will fix itself if the user changes their clock
-    var median_offset = Immutable.List(chainTimeOffset)
-        .sort()
-        .get(Math.floor((chainTimeOffset.length - 1) / 2));
-    // console.log("median_offset", median_offset)
-    return median_offset;
-};
-
-const addProposalData = (approvals, objectId) => {
-    let didImpact = false;
-    approvals.forEach(id => {
-        let impactedAccount = objectById.get(id);
-        if (impactedAccount && impactedAccount !== true) {
-            didImpact = true;
-            let proposals = impactedAccount.get("proposals", Immutable.Set());
-
-            if (!proposals.includes(objectId)) {
-                proposals = proposals.add(objectId);
-                impactedAccount = impactedAccount.set("proposals", proposals);
-                objectById.set(impactedAccount.get("id"), impactedAccount);
-            }
-        }
-    });
-    return didImpact;
-};
-
 function FetchChainObjects(method, object_ids, timeout, subMap) {
     let get_object = method.bind(chain_store);
 
@@ -1486,9 +1232,9 @@ const chain_store = {
     resetCache,
     setDispatchFrequency: dispatchFrequency.set,
     init,
-    _subTo,
+    //_subTo,
     unSubFrom,
-    _isSubbedTo,
+    //_isSubbedTo,
     onUpdate,
     notifySubscribers,
     subscribe,
@@ -1511,7 +1257,7 @@ const chain_store = {
     getAccountMemberStatus,
     getAccountBalance,
     fetchRecentHistory,
-    _updateObject,
+    //_updateObject,
     getObjectsByVoteIds,
     getObjectByVoteID,
     getHeadBlockDate,
