+++ lib/chain/src/ChainStore/store.js
@@ -1,4 +1,10 @@
-import {localVariable} from "./utils";
+export const localVariable = initVal => {
+    let local = initVal;
+    return {
+        get: () => local,
+        set: newVal => (local = newVal)
+    };
+};
 
 /** tracks everyone who wants to receive updates when the cache changes */
 export const subscribers = new Set();
