+++ lib/chain/src/ChainStore/utils.js
@@ -1,4 +1,9 @@
+import Immutable from "immutable";
 import ChainTypes from "../ChainTypes";
+import ee from "../EmitterInstance";
+import {objectById, headBlockTimeString, chainTimeOffset} from "./store";
+
+export const emitter = ee();
 
 const {object_type, impl_object_type} = ChainTypes;
 
@@ -46,10 +51,25 @@ export const timeStringToDate = timeString => {
     return new Date(timeString);
 };
 
-export const localVariable = initVal => {
-    let local = initVal;
-    return {
-        get: () => local,
-        set: newVal => (local = newVal)
-    };
+/** Clear an object from the cache to force it to be fetched again. This may
+ * be useful if a query failed the first time and the wallet has reason to believe
+ * it may succeede the second time.
+ */
+export const clearObjectCache = id => {
+    objectById.delete(id);
+};
+
+export const getHeadBlockDate = () => {
+    return timeStringToDate(headBlockTimeString.get());
+};
+
+export const getEstimatedChainTimeOffset = () => {
+    if (chainTimeOffset.length === 0) return 0;
+    // Immutable is fast, sorts numbers correctly, and leaves the original unmodified
+    // This will fix itself if the user changes their clock
+    var median_offset = Immutable.List(chainTimeOffset)
+        .sort()
+        .get(Math.floor((chainTimeOffset.length - 1) / 2));
+    // console.log("median_offset", median_offset)
+    return median_offset;
 };
